+++ HISTORY.md
@@ -1,3 +1,15 @@
+# 0.40.6
+
+Add an extra constructor for `LogDensityFunction(::Model, logdensity_function, ::AbstractTransformStrategy[, accs; adtype])`, which is a shorthand for generating the vectorised values according to the given transform strategy, and then calling the existing constructor with those values.
+
+This allows you to conveniently construct a linked `LogDensityFunction` in one line via
+
+```julia
+ldf = LogDensityFunction(model, logdensity_function, LinkAll())
+```
+
+instead of the old route where you had to create a `VarInfo` or `VectorValueAccumulator` first.
+
 # 0.40.5
 
 Remove DynamicPPL's internal variable transformation code and replace it directly with the new `Bijectors.VectorBijectors` module.